Full data retrieval library and updates to the code to adjust for the changes
